feat: xfer: BM-18026 + general improvments
Reindex calendar, todolist, addressbook after xfer
Persuades local "application data" of client applications (in browser)
Logout and Lockout (cyr_deny) user while doing a xfer: It's now somewhat safe to xfer while in production
Wait for the replication to be done before trying to xfer

+++ parent/mailbackend/net.bluemind.backend.mail.replica.service/src/net/bluemind/backend/mail/replica/service/internal/DbMailboxRecordsService.java
@@ -678,7 +678,9 @@ public class DbMailboxRecordsService extends BaseMailboxRecordsService implement
 		String folderUid = IMailReplicaUids.uniqueId(container.uid);
 		ItemValue<Mailbox> box = ServerSideServiceProvider.getProvider(SecurityContext.SYSTEM)
 				.instance(IMailboxes.class, container.domainUid).getComplete(container.owner);
-		indexService.deleteBox(box, folderUid);
+		if (box != null) {
+			indexService.deleteBox(box, folderUid);
+		}
 		storeService.prepareContainerDelete();
 	}
